Remove AAD Graph helpers, trim down MS Graph helpers to functions that are used by more than one service package
